+++ src/application.rs
@@ -5,15 +5,15 @@ use gtk::prelude::*;
 use gtk::subclass::prelude::*;
 use gtk::{gdk, gio, glib};
 
+use crate::color::color::Color;
 use crate::config::{APP_ID, PKGDATADIR, PROFILE, VERSION};
-use crate::model::color::Color;
 use crate::widgets::about_window::EyedropperAbout;
 use crate::widgets::preferences::PreferencesWindow;
 use crate::window::AppWindow;
 
 mod imp {
 
-    use crate::model::color::Color;
+    use crate::color::color::Color;
 
     use super::*;
     use adw::subclass::prelude::AdwApplicationImpl;